@rebel-jonghewk 2025-02-05
@mreso @agunapal 
+ Now vLLM's version is v0.7.1. Do you have plans to update or should I add a PR for it?